♻️ refactor: 함수명 변경 checkStart->winningResult
